+++ src/main/java/com/scanii/internal/DefaultScaniiClient.java
@@ -203,6 +203,21 @@ public class DefaultScaniiClient implements ScaniiClient {
   }
 
   @Override
+  public boolean deleteTrace(String id) {
+    Objects.requireNonNull(id, "resource id cannot be null");
+
+    HttpRequest req = newRequestBuilder(target.resolve("/v2.2/files/" + id + "/trace"))
+      .DELETE()
+      .build();
+    HttpResponse<String> response = send(req);
+
+    if (response.statusCode() != 204) {
+      parseAndThrowError(response);
+    }
+    return true;
+  }
+
+  @Override
   public Optional<ScaniiTraceResult> retrieveTrace(String id) {
     Objects.requireNonNull(id, "resource id cannot be null");